+++ Changelog.md
@@ -49,3 +49,5 @@ tsc no longer removes comments in generated code.  This can cause problems with 
 
 **1.2.7 / 2021-08-02**  
 Revert type declaration for AbstractConstructor which was broken during eslint integration.
+Update eslint related dev-dependencies.
+